refactor solver into smaller pieces
