+++ CMakeLists.txt
@@ -161,6 +161,9 @@ add_library(jres_solver_lib
   src/jres_internal_types.cpp
   src/jres_solver_base.cpp
   src/jres_standard_solver.cpp
+  src/analysis/capacity_analyzer.cpp
+  src/constraints/balancing.cpp
+  src/constraints/minimum_rest.cpp
 )
 set_target_properties(jres_solver_lib PROPERTIES OUTPUT_NAME "jres_solver")
