+++ src/jres_standard_solver.hpp
@@ -6,6 +6,7 @@
 #pragma once
 
 #include "jres_solver_base.hpp"
+#include "jres_internal_types.hpp"
 #include <memory>
 #include <set>
 #include <map>
@@ -29,45 +30,12 @@ private:
         std::map<std::pair<std::string, int>, int>& workVars
     );
 
-    // Incentivize balanced participation (Soft Constraint)
-    void add_balancing_constraints(
-        Highs &highs,
-        const std::vector<jres::internal::TeamMember> &participants,
-        const std::map<std::pair<std::string, int>, int>& workVars,
-        double avgStints
-    );
-
-    // Enforce minimum rest constraints (potentially across both roles)
-    void apply_minimum_rest_constraints(
-        Highs &highs,
-        const std::vector<jres::internal::TeamMember> &participants,
-        const std::map<std::pair<std::string, int>, int>& driverVars,
-        const std::map<std::pair<std::string, int>, int>& spotterVars,
-        bool enforceCombined
-    );
-
-    struct CapacityAnalysis {
-        int totalCapacity;
-        std::string details;
-    };
-
-    // Helper for pre-flight feasibility check
-    CapacityAnalysis calculate_max_potential_capacity(const std::vector<jres::internal::TeamMember>& participants);
-
-    struct SlackInfo {
-        std::string type;
-        std::string memberName;
-        int stintIndex;
-        double limit = 0.0;
-        double actual = 0.0;
-    };
-
     std::unique_ptr<Highs> m_highs;
     std::map<std::pair<std::string, int>, int> m_driverWorkVars;
     std::map<std::pair<std::string, int>, int> m_spotterWorkVars;
     std::map<std::pair<std::string, int>, int> m_switchVars;
 
     // Elastic Solver State
-    std::map<int, SlackInfo> m_slackInfo;
+    std::map<int, jres::internal::SlackInfo> m_slackInfo;
     std::set<int> m_unavailableVars;
 };
